test/dist-base.js split into 9 focused files
