+++ test/dist-base-special-cases.js
@@ -1,0 +1,165 @@
+import { assert } from 'chai'
+import { before, describe, it } from 'mocha'
+import { float } from '../src/core'
+import * as dist from '../src/dist'
+import PreComputed from '../src/dist/_pre-computed'
+
+describe('dist', () => {
+  describe('PreComputed', () => {
+    class PreComputedTestClass extends PreComputed {}
+    const preComputed = new PreComputedTestClass()
+
+    it('should throw error if _pk is not overridden', () => {
+      assert.throws(() => {
+        preComputed._pk()
+      }, 'PreComputed._pk() is not implemented')
+    })
+
+    it('should clamp _cdf to 1 on cache-hit path when PMFs sum to 1+epsilon', () => {
+      // Subclass whose PMFs sum to slightly above 1 (simulating floating-point rounding)
+      class OverflowPMF extends PreComputed {
+        _pk () { return 0.5 + 1e-15 }
+      }
+      const d = new OverflowPMF()
+      // Warm up cache by accessing index 1 (fills cdfTable[0] and cdfTable[1])
+      d._cdf(1)
+      // cdfTable[0] = 0.5+eps, cdfTable[1] ≈ 1 + 2eps > 1; re-reading from cache must still return ≤ 1
+      assert.isAtMost(d._cdf(0), 1)
+      assert.isAtMost(d._cdf(1), 1)
+    })
+
+    it('_generator returns NaN when all alias tables are exhausted (zero-probability PMF)', () => {
+      // _pk returns 0 for every k, so every alias table always routes to the overflow slot (TABLE_SIZE).
+      // After MAX_NUMBER_OF_TABLES tables the do-while exits without sampling, and must return NaN not undefined.
+      class ZeroMassDist extends PreComputed {
+        _pk () { return 0 }
+      }
+      const d = new ZeroMassDist()
+      assert(Number.isNaN(d._generator()))
+    })
+  })
+
+  // Degenerate is not covered by dist-cases.js (special-cased below) — verify constructor throws on missing params per issue #50.
+  describe('Degenerate', () => {
+    describe('constructor', () => {
+      it('should throw error if no parameters are provided', () => {
+        assert.throws(() => new dist.Degenerate())
+      })
+    })
+
+    describe('moments', () => {
+      it('mean should equal x0', () => {
+        assert.strictEqual(new dist.Degenerate(3).mean(), 3)
+        assert.strictEqual(new dist.Degenerate(-1.5).mean(), -1.5)
+      })
+
+      it('variance should be 0', () => {
+        assert.strictEqual(new dist.Degenerate(3).variance(), 0)
+      })
+
+      it('skewness should be NaN (0/0 at point mass)', () => {
+        assert(Number.isNaN(new dist.Degenerate(3).skewness()))
+      })
+
+      it('kurtosis should be NaN (0/0 at point mass)', () => {
+        assert(Number.isNaN(new dist.Degenerate(3).kurtosis()))
+      })
+    })
+
+    describe('.q()', () => {
+      it('quantile should return x0 for all probabilities', () => {
+        assert.strictEqual(new dist.Degenerate(5).q(0), 5)
+        assert.strictEqual(new dist.Degenerate(5).q(0.5), 5)
+        assert.strictEqual(new dist.Degenerate(5).q(1), 5)
+        assert.strictEqual(new dist.Degenerate(-2).q(0.1), -2)
+        assert.strictEqual(new dist.Degenerate(0).q(0.99), 0)
+      })
+    })
+  })
+
+  // Kolmogorov: open lower boundary — x=0 is outside the support (x>0) and must return 0.
+  describe('Kolmogorov', () => {
+    const k = new dist.Kolmogorov()
+
+    describe('.pdf()', () => {
+      it('should return 0 at the open lower boundary x = 0', () => {
+        assert.equal(k.pdf(0), 0)
+      })
+    })
+
+    describe('.cdf()', () => {
+      it('should return 0 at the open lower boundary x = 0', () => {
+        assert.equal(k.cdf(0), 0)
+      })
+    })
+
+    describe('.survival()', () => {
+      it('should return 1 at the open lower boundary x = 0', () => {
+        assert.equal(k.survival(0), 1)
+      })
+    })
+  })
+
+  // Base class helper: _qEstimateWalk.
+  describe('Distribution._qEstimateWalk', () => {
+    // Poisson(5) reference values:
+    //   ppf(0.1) = 2   cdf(2)=0.12465  cdf(1)=0.04043
+    //   ppf(0.5) = 5   cdf(5)=0.61596  cdf(4)=0.44049
+    //   ppf(0.9) = 8   cdf(8)=0.93191  cdf(7)=0.86663
+    let d
+    before(() => { d = new dist.Poisson(5) })
+
+    it('should walk up from a start below the quantile', () => {
+      assert.strictEqual(d._qEstimateWalk(0.9, 0), 8)
+    })
+
+    it('should walk down from a start above the quantile', () => {
+      assert.strictEqual(d._qEstimateWalk(0.1, 20), 2)
+    })
+
+    it('should return immediately when start already satisfies the invariant', () => {
+      assert.strictEqual(d._qEstimateWalk(0.5, 5), 5)
+    })
+
+    it('should walk up from a negative start', () => {
+      assert.strictEqual(d._qEstimateWalk(0.5, -10), 5)
+    })
+
+    it('should return lower support boundary when p=0', () => {
+      assert.strictEqual(d._qEstimateWalk(0, 5), 0)
+    })
+
+    it('should return upper support boundary when p=1', () => {
+      assert.strictEqual(d._qEstimateWalk(1, 5), Infinity)
+    })
+  })
+
+  // Degenerate distribution.
+  describe('Degenerate', () => {
+    describe('.sample()', () => {
+      describe('random parameters', () => {
+        it('should generate values with Degenerate distribution', () => {
+          const x0 = float()
+          const degenerate = new dist.Degenerate(x0)
+          degenerate.sample(10).forEach(d => {
+            assert(d === x0)
+          })
+        })
+      })
+    })
+
+    describe('.pdf(), .cdf()', () => {
+      describe('random parameters', () => {
+        it('differentiating cdf should give pdf', () => {
+          const x0 = float()
+          const degenerate = new dist.Degenerate(x0)
+          assert.equal(degenerate.pdf(x0), 1)
+          assert.equal(degenerate.pdf(x0 + 1), 0) // fixed offset guarantees argument != x0
+          assert.equal(degenerate.cdf(x0 - 1), 0) // fixed offset guarantees argument < x0
+          assert.equal(degenerate.cdf(x0), 1)
+          assert.equal(degenerate.cdf(x0 + Math.random()), 1)
+        })
+      })
+    })
+  })
+})
